RaspberryPi : Download directory should be stored external USB device
As in sd card, the space will be not enough to handle files download.

Prefer to handle this, to put /Download directory into an external device

Menu nas display a message if /Download is not mounted on USB device.

Modify udev mount script to automount the first usb key plugged to the directory /Download

Signed-off-by: Baptiste DURAND <baptiste.durand@gmail.com>